Moved main screen logic into separate screen file. Created package for all app screens. Updated main activity to contain the scaffold and load that with the home screen
