+++ app/src/main/java/com/example/monthlyspendingtracker/MainActivity.kt
@@ -32,6 +32,7 @@ import java.util.Calendar
 import java.util.Currency
 import java.util.Date
 import androidx.compose.material.ModalDrawer
+import com.example.monthlyspendingtracker.screens.HomeScreen
 
 //enum class Category(val label: String, val icon: Unit) {
 //    DANIEL_FUN("Daniel Fun", Icons.Default.Money),
@@ -72,186 +73,10 @@ fun DefaultPreview() {
     }
 }
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun TopBar(onMenuClick: () -> Unit) {
-    TopAppBar(
-        title = { Text(text = "Monthly Spending Tracker") },
-        navigationIcon = {
-            IconButton(onClick = onMenuClick) {
-                Icon(painterResource(R.drawable.baseline_menu_24), contentDescription = null)
-            }
-        },
-        // background color of topAppBar (cannot find "backgroundColor" -- not sure what to do here)
-//        backgroundColor = Color(0xFF0F9D58)
-    )
-}
-
-@Composable
-fun Drawer() {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-        repeat(5) { item ->
-            Text(text = "Item $item", color = Color.Black)
-        }
-    }
-}
-
-@Composable
-fun SpendingTrackerScaffold(
-    content: @Composable () -> Unit
-) {
-    val drawerState =
-        androidx.compose.material.rememberDrawerState(androidx.compose.material.DrawerValue.Closed)
-    val scope = rememberCoroutineScope()
-
-    ModalDrawer(
-        drawerState = drawerState,
-        gesturesEnabled = drawerState.isOpen,
-        drawerContent = {Drawer()},
-        content = {
-            Column {
-                TopBar(
-                    onMenuClick = {
-                        scope.launch {
-                            if (drawerState.isOpen) {
-                                drawerState.close()
-                            } else {
-                                drawerState.open()
-                            }
-                        }
-                    }
-                )
-                content()
-            }
-        }
-    )
-}
-
 @ExperimentalComposeUiApi
 @Composable
 fun MyApp() {
-    val database = Room.databaseBuilder(
-        LocalContext.current,
-        AppDatabase::class.java,
-        "expenses-db"
-    ).build()
-
-    // Initialize the running total with expenses from the current month
-    val currentMonth = Calendar.getInstance().apply {
-        set(Calendar.DAY_OF_MONTH, 1) // Set to the first day of the month
-    }.time
-
-    var totalAmount by remember { mutableDoubleStateOf(0.0) }
-
-    LaunchedEffect(Unit) {
-        val amountFromDb = withContext(Dispatchers.IO) {
-            database.expenseDao().getTotalAmountForMonth(currentMonth) ?: 0.0
-        }
-        totalAmount = amountFromDb
-    }
-
-    var purchaseAmount by remember { mutableStateOf("") }
-    var selectedCategory by remember { mutableStateOf(Category.DANIEL_FUN) }
-    val keyboardController = LocalSoftwareKeyboardController.current
-
-    val format: NumberFormat = NumberFormat.getCurrencyInstance()
-    format.maximumFractionDigits = 2
-    format.currency = Currency.getInstance("USD")
-
-    SpendingTrackerScaffold {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(16.dp),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Center
-        ) {
-            // Display running total
-            Text(
-                text = "${format.format(totalAmount)}",
-                style = MaterialTheme.typography.titleMedium,
-                modifier = Modifier.padding(16.dp)
-            )
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-            // Text box to enter purchase amount
-            OutlinedTextField(
-                value = purchaseAmount,
-                onValueChange = { purchaseAmount = it },
-                label = { Text("Enter purchase price (USD)") },
-                keyboardOptions = KeyboardOptions.Default.copy(keyboardType = KeyboardType.Number),
-                leadingIcon = { Icon(painterResource(R.drawable.ic_baseline_attach_money_24), contentDescription = null) },
-                modifier = Modifier.fillMaxWidth(),
-                colors = OutlinedTextFieldDefaults.colors(focusedTextColor = Color.Black, unfocusedTextColor = Color.Black)
-            )
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-            // Dropdown menu for categories
-            var expanded by remember { mutableStateOf(false) }
-            OutlinedTextField(
-                value = selectedCategory.label,
-                onValueChange = {},
-                label = { Text("Category") },
-                trailingIcon = {
-                    Icon(
-                        painterResource(R.drawable.ic_baseline_keyboard_arrow_down_24),
-                        contentDescription = null,
-                        modifier = Modifier.clickable { expanded = true }
-                    )
-                },
-                readOnly = true,
-                keyboardOptions = KeyboardOptions.Default.copy(keyboardType = KeyboardType.Text),
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .onFocusChanged { if (it.isFocused) expanded = true },
-                colors = OutlinedTextFieldDefaults.colors(focusedTextColor = Color.Black, unfocusedTextColor = Color.Black)
-            )
-
-            DropdownMenu(
-                expanded = expanded,
-                onDismissRequest = { expanded = false },
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                for (category in Category.values()) {
-                    DropdownMenuItem(
-                        text= { Text(category.label) },
-                        onClick = {
-                            selectedCategory = category
-                            expanded = false
-                        },
-                    )
-                }
-            }
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-            // Submit button
-            IconButton(
-                onClick = {
-                    val amount = purchaseAmount.toDoubleOrNull() ?: 0.0
-                    totalAmount += amount
-
-                    // Create an Expense object and insert it into the database
-                    val expense = ExpenseEntity(date = Date(), category = selectedCategory.label, price = amount)
-                    CoroutineScope(Dispatchers.IO).launch {
-                        database.expenseDao().insertExpense(expense)
-                    }
-
-                    purchaseAmount = ""
-                    selectedCategory = Category.DANIEL_FUN
-                    keyboardController?.hide()
-                }
-            ) {
-                Icon(painterResource(R.drawable.ic_baseline_add_24), contentDescription = null)
-            }
-        }
+    SpendingTrackerScaffold{
+        HomeScreen()
     }
 }
